% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_short_summaries.R
\name{combineDFComp}
\alias{combineDFComp}
\title{Combine short summary files from exported summaries}
\usage{
combineDFComp(path = ".", filename = "001_comparedf_summary", open = TRUE)
}
\arguments{
\item{path}{\code{character} defining the path to the files that needs to be summarised}

\item{filename}{\code{character} defining the name of the summary html output}

\item{open}{Should the generated summary file be opened}
}
\value{
creates an html table with combined summary info
}
\description{
Combine short summary files from exported summaries
}
