% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.summary.comparedf.R
\name{export.summary.comparedf}
\alias{export.summary.comparedf}
\title{The export method for a \code{summary.comparedf} object}
\usage{
\method{export}{summary.comparedf}(
  x,
  ...,
  type = c("html", "word", "pdf", "github"),
  db = NULL,
  folder = "parprog",
  subfolder = "",
  filename = "",
  title = "",
  author = Sys.info()[["user"]],
  deleteRmd = TRUE
)
}
\arguments{
\item{x}{An object returned by the \code{summary.comparedf} function.}

\item{...}{Other arguments passed to \code{\link[rmarkdown]{render}}. In
\code{print}, these are passed to \code{\link[knitr]{kable}}.}

\item{type}{Type of output - html, word or pdf.}

\item{db}{An \code{\link[NNaccess]{nnaccess}} object}

\item{folder}{The folder in the instance - default: parprog, but it could
also be nonctrprog.}

\item{subfolder}{Optional subfolder to folder.}

\item{filename}{Name of file without extension, when omitted it defaults to
the name of the script that executes the command.}

\item{title}{Title in the header of the document.}

\item{author}{Author of the document - default: the user that executes the
program.}

\item{deleteRmd}{If TRUE the Rmd file will be deleted, else not.}
}
\description{
Exports the output of the \code{\link{summary.comparedf}} object to a file.
}
\seealso{
\code{\link{summary.comparedf}}, \code{\link{comparedf.control}}
}
