% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngCompare.R
\name{plot.pngcomp}
\alias{plot.pngcomp}
\title{Plot a comparison of two png files}
\usage{
\method{plot}{pngcomp}(
  x,
  ...,
  base = c("png1", "png2"),
  higlight_change = TRUE,
  show_foot_line = FALSE,
  errorcol = 2,
  okcol = 3,
  lwd = 4,
  debug = FALSE
)
}
\arguments{
\item{x}{An object created by pngCompare}

\item{...}{Additional arguments currently ignored}

\item{base}{Which of the plots should be considered the base plot, png1
corresponds to file1 and png2 corresponds to file2}

\item{higlight_change}{Should the changes be hihglighted}

\item{show_foot_line}{Should the line separating the system footnote from the rest of the graph be shown}

\item{errorcol}{colour of the error box}

\item{okcol}{colour of the systemfootnote box}

\item{lwd}{line width of the box, measured in number of pixels}

\item{debug}{\code{logical} should a debug plot be shown}
}
\value{
x is returned invisibly.
}
\description{
Plot a comparison of two png files
}
