% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.comparedf.R
\name{print.summary.comparedf}
\alias{print.summary.comparedf}
\title{print.summary.comparedf
The summary method for a \code{comparedf} object

Print a more detailed output of the \code{\link{comparedf}} object.}
\usage{
\method{print}{summary.comparedf}(x, ..., format = "pandoc")
}
\arguments{
\item{x}{An object of class \code{"summary.comparedf"}, as made by the \code{\link{comparedf}} S3 method.}

\item{...}{Other arguments passed to \code{\link{print.summary.comparedf}}. In \code{print}, these are passed to \code{\link[knitr]{kable}}.}

\item{format}{Passed to \code{\link[knitr]{kable}}: the format for the table. The default here is "pandoc".
To use the default in \code{kable}, pass \code{NULL}.}
}
\value{
An object of class \code{"summary.comparedf"} is returned.
}
\description{
print.summary.comparedf
The summary method for a \code{comparedf} object

Print a more detailed output of the \code{\link{comparedf}} object.
}
