% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngCompare.R
\name{print.summary.pngcomp}
\alias{print.summary.pngcomp}
\title{Print a summary of a comparison between png files}
\usage{
\method{print}{summary.pngcomp}(x, ...)
}
\arguments{
\item{x}{The comparison to print}

\item{...}{Additional arguments passed on, not currently in use.
export}
}
\description{
Print a summary of a comparison between png files
}
