% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{comparedf}
\alias{diffs}
\alias{n.diffs}
\alias{n.diff.obs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{arsenal}{\code{\link[arsenal]{comparedf}}, \code{\link[arsenal]{diffs}}, \code{\link[arsenal:diffs]{n.diff.obs}}, \code{\link[arsenal:diffs]{n.diffs}}}
}}

